# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Cardano Node API",
        "title": "cardano-node-api",
        "contact": {
            "name": "Blink Labs",
            "url": "https://blinklabs.io",
            "email": "support@blinklabs.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost",
    "basePath": "/api",
    "paths": {
        "/localtxmonitor/has_tx/{tx_hash}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "localtxmonitor"
                ],
                "summary": "Check if a particular TX exists in the mempool",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.responseLocalTxMonitorHasTx"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.responseApiError"
                        }
                    }
                }
            }
        },
        "/localtxmonitor/sizes": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "localtxmonitor"
                ],
                "summary": "Get mempool capacity, size, and TX count",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.responseLocalTxMonitorSizes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.responseApiError"
                        }
                    }
                }
            }
        },
        "/localtxmonitor/txs": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "localtxmonitor"
                ],
                "summary": "List all transactions in the mempool",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.responseLocalTxMonitorTxs"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.responseApiError"
                        }
                    }
                }
            }
        },
        "/localtxsubmission/tx": {
            "post": {
                "description": "Submit an already serialized transaction to the network.",
                "produces": [
                    "application/json"
                ],
                "summary": "Submit Tx",
                "parameters": [
                    {
                        "enum": [
                            "application/cbor"
                        ],
                        "type": "string",
                        "description": "Content type",
                        "name": "Content-Type",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.responseApiError": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
        "api.responseLocalTxMonitorHasTx": {
            "type": "object",
            "properties": {
                "has_tx": {
                    "type": "boolean"
                }
            }
        },
        "api.responseLocalTxMonitorSizes": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "tx_count": {
                    "type": "integer"
                }
            }
        },
        "api.responseLocalTxMonitorTxs": {
            "type": "object",
            "properties": {
                "tx_bytes": {
                    "type": "string",
                    "format": "base64",
                    "example": "\u003cbase64 encoded transaction bytes\u003e"
                },
                "tx_hash": {
                    "type": "string",
                    "format": "base16",
                    "example": "96649a8b827a5a4d508cd4e98cd88832482f7b884d507a49466d1fb8c4b14978"
                }
            }
        }
    }
}